[new-eval] New evaluation split in different phases
The new evaluation is composed of 5 phases:
 - init phase;
 - unload phase;
 - backpressure phase;
 - targeting phase;
 - averaging phase.

Every phases launches a set of batches (measured in seconds,
configurable).

During "init" the system sends two batches of 5s (configurable) one
record at a time in order to set connections and init the system.

During "unload" the system sends records one at a time in batches and
averages latency results in order to calculate the unloaded latency.

During "backpressure" the system sends 1 batch at maximum rate and
measures the throughput. We use the natural Flink backpressure mechanism
to establish a starting point to estimate a good throughput. This should
be an estimate of the maximum throughput at which the system can run.
This phase is required in order to start with an input rate that should
be closer to the sustainable throughput than starting from a super-low
value and increase. This practice shortens the experiment.

During "targeting", the system uses the "backpressure" throughput and
calculates the latency every second and it increases the input rate
until the measured latency is bigger than some latency threshold. This
means that we reached an input rate around which the system starts
enqueuing.  The latency threshold is the time that a record should wait
in queue if N records are before it. It is calculated by using the
latency unloaded times the number of records we tolerate to be in queue
(configurable).

During "averaging" the system uses the "targeting" throughput and
fluctuates around it by increasing the rate when the latency gets below
the target latency and decreases it otherwise. This behavior guarantees
that you are running at the stable throughput: if you are able to
fluctuate around a fixed latency it means that you are either wasting
time idle, nor filling queues; i.e. this is the maximum workload that
you can handle without growing queues undefinitely.

The final results are the unloaded latency and the mean throughput
calculated during "averaging".

* Adaptation of plotting and parsing scripts to new result format